Set up ApiService and create first resource, AuthResource
Install axios for HTTP requests. axios is async/await compatible out of
the box and allows for easy mocking, along with axios-mock-adapter and
ts-sinon for mocking during testing.

Sets up the API service to make standardized API requests, and sets up
the first public methods on the client via AuthResource, which contains
a method isSessionValid

Issue #2 - Method to validate session