Add example script for getting a folder share link
This draws heavily on @andrewatwood's existing upload example script
and the code he wrote in the README to explain getting share links for
newly created records and folders.  I still find it useful to have a
full test script even with these other resources.  You can check the
returned value against the share link shown in the web client under
the Sharing tab on the right side.  If you click "Share" you'll see
the share link in the modal.