WIP: Expose getRoot through the SDK
This was already partially implemented, confusingly.  Clients of the
SDK should be able to get the user's default archive from the user's
session, just as the web client does, without needing to pass in an
archive number with each request.

The next step is to always use the session user's default archive
instead of requiring the archive number to be passed by the SDK (that
is, to get the default archive on init).  Most likely I will keep the
option to pass in the archive number for backwards compatibility.

Also added another example script here to demonstrate usage and test
my work, though these are rapidly becoming repetitive and less useful.