Get account and default archive from session
Given a session in the SDK, navigate to the given archive if one was
entered or the default archive for the logged-in user if not.

As currently written, this requires some back-end changes as well.

Evidently the `archiveNbr` in the `folder` table does not correspond
to the `archiveNbr` in the `archive` table.  One `archiveId` in
`folder` has multiple `archiveNbr` values.  This is strange, but in
any case it means that getRoot will not return the necessary
archiveNbr to set the default archive.