Add storage based on record size
This is only partly functional, because the record does not instantly
have a size (it has to go through processing first).  I need to find
out how best to manage that.

It might be irrelevant because the next thing I need to do is switch
to the new upload flow.  I might get the size in that flow
pre-processing, or at least there might be an obvious way to wait for
it before adding the storage.

We also intend to only add storage if the upload succeeded, which is
not implemented here yet.

HOWEVER, given a record size this now works to add storage to an
account.

It also needs the pad secret we have discussed, so that we aren't
exposing an endpoint that allocates an arbitrary amount of storage.
That is a change to make here and in the backend.