Save archive id to session and use RegisterRecord
RegisterRecord should now work and will take a single object, not an array.

When the team wrote RegisterRecord, they expected it to only take one
object at a time.  This makes it different from our other endpoints,
which expect an array of endpoints.  Thanks to @jasonaowen for the
pointer about using pipe to allow either option from the SDK.

Also, add more detail to the example upload script.