Improve type management
Make recordId optional in RecordVO.  We use RecordVO
before it exists in the database, so it will not yet
have a RecordId.  Thanks for the catch, @jasonaowen!  However, the
issue with this is that a RecordVO really shouldn't exist without
an archiveId.  We need to make it optional in the definition to match
up with the fact that we might fail to get an archive ID elsewhere.

Also return the new, more specific, PresignedUrlResponse type from
`getPresignedUrl` instead of a SimpleVO.