Rewrite the upload script
This is what I have been using to test the
`registerRecordAndAddStorage` function.  I'm not sure it's actually
useful for anyone else, but it's the best answer to @jasonaowen's
comment about the previous script being out of date.  This commit
shows a little bit of web-app code, which isn't open source yet, but I
think it's fine.