Remove Permanent API key
The Permanent API no longer requires an API key. Requests don't need to
include it, no behavior is gated on it, and soon the API won't even know
what an API key is. Stop requiring the key during initialization, and
stop sending it with requests.

At least until the scheduled auth work is finished, though, we would
like potential users of the SDK to talk to us about gaining access and
configuring the SDK. That conversation will no longer include API keys,
so clarify the instructions to contact us.

PER-8278 Rotate or delete Permanent API key