Remove dependency on cz-conventional-changelog
As far as I can tell, we do not and have not used
cz-conventional-changelog or the Commitizen ecosystem it depends on in
turn. We can have a discussion about whether or not we want to, but in
the meantime, let's remove the outdated version that depends on
vulnerable versions of other packages.