Remove dependency on standard-version
We have not used the `version` command, which depends on the
standard-version package[1]. The version of standard-version we had
depended on vulnerable versions of several other packages. We can decide
together later if we want to bring this back in, but for now, fix the
security issues by removing the dependency.