Refactor API Service construction from cookies
Before we can meaningfully add token authentication, we need to refactor
the way the ApiService is built to allow for an alternate form of
construction. Make the constructor private, and add a new static factory
method that contains the cookie-specific configuration. While doing so,
also inline the previously separate function that creates the `Cookie`
header.

PER-8535 Use OAuth in node-sdk