Use Joi to validate client configuration
We were previously using simple if statements to validate that all
required parameters are passed in. That worked well enough when there
are only two required parameters, but it gets more complicated with more
complex requirements.

Allowing token authentication side-by-side with session authentication
will require two mutually exclusive sets of parameters. Joi[1] is a
library that handles validation, and is a transitive dependency of
another library that will be used as part of the OAuth work. Use it
directly to validate that both cookies required for session
authentication are present.

Note that this could be construed as a breaking change: the thrown error
type for constructing a `Permanent` instance with incorrect parameters
changes from `PermSdkError` to a Joi `ValidationError`. I also learned
that Nodejs does not support the `options.cause` argument to the Error
constructor[2], so we cannot even extend `PermSdkError` to preserve the
original; we would have to just flatten it into a string. Given that
constraint, I felt that forcing the error Joi throws into a
`PermSdkError` would be less useful, as it would obscure the detailed
validation information Joi can provide.

[1] https://joi.dev/
[2] https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Error#browser_compatibility

PER-8535 Use OAuth in node-sdk