Add OAuth client
Add a way for authenticating to the Permanent API via OAuth. This is
critical functionality for our partners, and particularly for our
Etherpad plugin.

As before, some private configuration is required: clients need an OAuth
client ID and client secret, which can only be issued by Permanent. Once
that's done, they can initiate an Authorization Code Grant login
flow[1]. When the user returns to their application, they can then
complete the flow to get an instance of the `Permanent` client that is
configured with the token.

Behind the scenes, we are using simple-oauth2 (npm[2], repo[3]) to
handle the OAuth workflow.

This new functionality should work alongside the legacy session-based
authentication, and so should not break any existing clients.

[1] https://fusionauth.io/docs/v1/tech/oauth/#example-authorization-code-grant
[2] https://www.npmjs.com/package/simple-oauth2
[3] https://github.com/lelylan/simple-oauth2

PER-8535 Use OAuth in node-sdk

Co-authored-by: Cecilia Krum <cecilia@permanent.org>