Check login status before getting session info
The API currently requires a CSRF token on many endpoints. In order for
a newly constructed instance to be able to talk to the API, it first
needs to call an endpoint that does not require a CSRF token but does
return one in its response, which can then be used in future requests.

The `useArchive` workflow, which would happen in the legacy session auth
if you specified an `archiveNbr` before calling `init()`, does this: it
first calls `/api/auth/loggedIn` to ensure the credentials are valid,
and then makes the actual requests necessary using the newly-returned
CSRF token.

However, if the user did not specify an `archiveNbr` in the constructor,
the `init()` called `getAccountArchive()`, which did not so validate the
credentials. As a result, the CSRF store was not initialized, and
getting the session info from `/api/account/getSessionAccount` failed
due to a lack of CSRF token.

Validate the given credentials as part of initialization in order to
populate the CSRF token.