Remove codecov dependency and script
Remove the `cov:send` npm script, and the dependency on codecov[1] which
provided its binary. The codecov binary is a tool to upload coverage
reports to Codecov.io.

It looks like the GitHub repo does have a secret configured,
`CODECOVERAGEGITHUBTOKEN`, and there is an empty repository in
Codecov[2]. However, as far as I can tell, we have never uploaded any
coverage reports to it, and it is not part of our CI.

The codecov package is deprecated[3]. The version we have also has
insecure transitive dependencies.

Rather than figure out how to replace it, simply delete it.

[1] https://www.npmjs.com/package/codecov
[2] https://app.codecov.io/gh/PermanentOrg/node-sdk/
[3] https://about.codecov.io/blog/codecov-uploader-deprecation-plan/