Remove dependency on nyc
The nyc project[1] builds code coverage reports for our unit tests.
However, it appears not to be maintained any longer, and it has
transitive dependencies that have critical security vulnerabilities.

We're not actively working on this repo any more, and we never had fully
configured code coverage reports, so the easiest solution is to simply
remove our dependency on nyc.

[1] https://www.npmjs.com/package/nyc