Fix license in package.json
The `package.json` shows a license of "MIT" but the `LICENSE` file shows AGPLv3.

My recollection is that the license was meant to be AGPLv3, so MIT is a typo.
This commit replace "MIT" with the SPDX identifier for AGPLv3.

This bug was committed when Andrew created the repo four years ago.  Both the
`package.json` and the `LICENSE` were part of the very first commit. Does that
mean I win the prize for fixing the oldest bug in the codebase?